+++ src/app/app.component.query.spec.ts
@@ -1,0 +1,5 @@
+export class AppComponentQuery {
+  public static getToolbar(parentElem: HTMLElement): HTMLElement | null {
+    return parentElem.querySelector('app-toolbar');
+  }
+}