+++ src/app/app.module.ts
@@ -1,14 +1,15 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
+import { CoreModule } from '@core/core.module';
+import { LayoutModule } from '@layout/layout.module';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { CoreModule } from '@core/core.module';
 
 @NgModule({
+  imports: [BrowserModule, BrowserAnimationsModule, CoreModule, LayoutModule, AppRoutingModule],
   declarations: [AppComponent],
-  imports: [BrowserModule, AppRoutingModule, BrowserAnimationsModule, CoreModule],
   providers: [],
   bootstrap: [AppComponent]
 })