+++ src/app/core/core.module.ts
@@ -1,10 +1,8 @@
 import { NgModule } from '@angular/core';
 
-import { LayoutModule } from './layout/layout.module';
-
 @NgModule({
-  imports: [LayoutModule],
-  exports: [LayoutModule],
+  imports: [],
+  exports: [],
   declarations: [],
   providers: []
 })