+++ src/app/shared/services/dialog/dialog.service.spec.ts
@@ -1,7 +1,7 @@
 import { TestBed } from '@angular/core/testing';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
-import { ServicesModule } from './services.module';
+import { DialogModule } from './dialog.module';
 import { DialogService } from './dialog.service';
 import { librariesMock } from '@core/data/libraries.mock';
 import { LibraryDialogComponent } from '@shared/components/library-dialog/library-dialog.component';
@@ -11,7 +11,7 @@ describe('DialogService', () => {
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [BrowserAnimationsModule, ServicesModule]
+      imports: [BrowserAnimationsModule, DialogModule]
     });
     service = TestBed.inject(DialogService);
   });